requirements.txt
langchain>=0.2.11,<0.3
langchain-community>=0.2.11,<0.3
langchain-openai>=0.1.16,<0.2
langchain-text-splitters>=0.2.2,<0.3
langchain-chroma>=0.1.1,<0.2
chromadb>=0.5.4,<0.6
pypdf
gradio>=4.44,<5
python-dotenv
tiktoken